# DOCS/API/openapi.yaml
openapi: 3.0.3
info:
  title: Najika World Complete API
  description: |
    Complete backend API for Najika World ecosystem.

    ## Features
    - 🔐 JWT Authentication & Security
    - 🎮 Game Systems (Combat, Inventory, Quests, Progression)
    - 🤖 AI Training (LoRA, Sessions, Code, Voice, Personality)
    - 🎤 Voice (Whisper AI, Edge TTS, WebSocket Streaming)
    - 👥 Multiplayer (Real-time WebSocket, Rooms, Player Sync)
    - ⚙️ Admin (User Management, System Monitoring, Logs)
    - 📊 Analytics & Monitoring (User tracking, Performance)
    - 🔒 Security (Encryption, TOR Integration)

    ## Base URL
    - Development: `http://localhost:8000`
    - Production: `https://api.najika.world`

    ## Authentication
    Most endpoints require JWT Bearer token authentication.
    Obtain a token via `/api/auth/login` or `/api/auth/register`.

  version: 5.0.0
  contact:
    name: Najika World Development Team
    url: https://najika.world
    email: support@najika.world
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.najika.world
    description: Production server
...
paths:
...
  /api/voice/transcribe:
    post:
      tags: [Voice]
      summary: Transcribe audio (Whisper AI)
      description: Convert audio to text using OpenAI Whisper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [audio_base64]
              properties:
                audio_base64:
                  type: string
                  description: Base64 encoded audio (WAV/MP3)
                audio_format:
                  type: string
                  enum: [wav, mp3, ogg]
                  default: wav
                language:
                  type: string
                  example: en
      responses:
        '200':
          description: Transcription result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  text:
                    type: string
                    example: Hello, this is a test.
                  language:
                    type: string
                    example: en
                  confidence:
                    type: number
  /api/voice/tts:
    post:
      tags: [Voice]
      summary: Text-to-Speech (Edge TTS)
      description: Convert text to speech audio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  example: EXPLOSION!
                language:
                  type: string
                  enum: [en, de, ja, es, fr]
                  default: en
                voice:
                  type: string
                  example: en-US-JennyNeural
                rate:
                  type: string
                  example: "+0%"
                personality:
                  type: string
                  enum: [megumin, aqua, darkness, kazuma]
      responses:
        '200':
          description: Generated audio
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  audio_base64:
                    type: string
                  format:
                    type: string
                    example: mp3
                  cached:
                    type: boolean

  # ============================================================================
  # MULTIPLAYER
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
...